build off of pseudorandom generator
using random signal kept getting me stuck based on querying the results
of the random signal board. I can't figure out how to not get a signal
of signals in that case. The generator library does the job for now by
creating a pure function with a seed. Possibly can use ports to generate
an external input which could build up a set of boards from a solvable 0
point.